feat: 참여자 이름 원시값 포장
